+++ src/components/shared/ResizableContainer.vue
@@ -1,15 +1,13 @@
 <template>
-  <div>
-    <div :style="style" ref="resizedWrapper" class="position--relative">
-      <slot></slot>
+  <div :style="style" ref="resizedWrapper" class="position--relative">
+    <slot></slot>
 
-      <v-btn flat
-             class="v-btn-absolute-bottom-right"
-             title="Resize"
-             @mousedown="onDragStart">
-        <v-icon>vertical_align_center</v-icon>
-      </v-btn>
-    </div>
+    <v-btn flat
+           class="v-btn-absolute-bottom-right"
+           title="Resize"
+           @mousedown="onDragStart">
+      <v-icon>vertical_align_center</v-icon>
+    </v-btn>
   </div>
 </template>
 
@@ -48,7 +46,8 @@
     },
     methods: {
       triggerResize () {
-        if (window !== undefined) window.dispatchEvent(new Event('resize'))
+        if (typeof window === 'undefined') return
+        window.dispatchEvent(new Event('resize'))
       },
       onDragStart (e) {
         this.resizing = true
@@ -65,23 +64,21 @@
         this.resizing = false
         this.dragStartY = 0
         this.$nextTick(() => {
-          if (window !== undefined) {
+          if (typeof window !== 'undefined') {
             this.triggerResize()
           }
         })
       }
     },
     mounted () {
-      if (window !== undefined) {
-        window.addEventListener('mouseup', this.onDragEnd)
-        window.addEventListener('mousemove', this.onDrag)
-      }
+      if (typeof window === 'undefined') return
+      window.addEventListener('mouseup', this.onDragEnd)
+      window.addEventListener('mousemove', this.onDrag)
     },
     destroyed () {
-      if (window !== undefined) {
-        window.removeEventListener('mouseup', this.onDragEnd)
-        window.removeEventListener('mousemove', this.onDrag)
-      }
+      if (typeof window === 'undefined') return
+      window.removeEventListener('mouseup', this.onDragEnd)
+      window.removeEventListener('mousemove', this.onDrag)
     },
     components: {
       BtnGroup